Use pref system directly to get flash status and clean up code
We now use the pref "plugin.state.flash" directly to get and set the status of the
flash plugin and to listen for changes. This is necessary if we want to support
click_to_play in the future, because although the PluginHost api supports it,
the AddonManager, which is used for change notifications, does not support it.

The new code was bundled in the module PluginStateManager. Therefore the existing
very ugly code using the PluginHost could be completely removed.

This results in a overall code reduction in the main module from 228 to 82 lines,
which hopefully will make maintenance and new features much easier in the future.