Add icon to show the 'ask to activate' (click_to_play) state
Until now click_to_play was shown as enabled.
This commit adds a new icon for the click_to_play that is derived from the
icon for the enabled state and shows it when this mode is selected.

Note that this commit does not allow you to use the addon to switch the flash
plugin to click_to_play. For this see the following commits.