Fix button displayed 2px too large in firefox versions below 40
In firefox-version before 40, sdk-buttons that use svg images are displayed too
large, namely 18x18px instead of 16x16px. This results in an navigation toolbar
height increase of 2 px, when the addon is installed.
To work around this in older versions of firefox, we simply set the width and
height attribute of the <svg> to 16 (px), which results in the button beeing
displayed in the correct size.